requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
google-generativeai
python-dotenv
gunicorn
waitress
